Mode is now used by split()
Prev split was called without mode, and defaulted to train always.

It now parses the mode and passes it to split() in main()

Beam file can potentially be an empty file with 1 header line.

To check for this case, open the beam file and count lines.

If there are at least two lines in the file, append it to the file_list.

If the file has less than two lines, skip and don't add.