Fix logic error and typo, comment out numbers var.
If largest_bin = len(hist[0]) -1, the largest bin is the last hist bin.

In that case the slice can only go to largest_bin +2 not +3.

hist[1][largest_bin +1] is the left edge of the largest bin.

To include it in the slice you have to add one to the index.

The slice should be to largest_bin +2 not +3.

largest_bin +3 only works if there is a bin after the largest bin.

Change second elif largest_bin ==1 to ==0.

Now there's a check for when the largest bin is the first histogram bin